requirements.txt
autopep8
langchain
python-dotenv
gunicorn
uvicorn
openai
unstructured
supabase
pinecone-client[grpc]